Adds cache for session relevant messages
adds a cache of messages to each session. The cache is cleaned
whenever all the subjects in a session proceed past the current
period. This reduces the cost of syncs: instead of unmarshalling
the entire message digest from Redis and searching it for
relevant messages to send, a relatively small cache of
relevant messages is maintained on the router.